+++ CHANGELOG.md
@@ -6,3 +6,5 @@
    - 
 - Fixes/updates
    - Standardizes output directory structure and parameter configuration in a JSON file
+   - Adds a new process.py script to parse fmwork output and write metrics to a CSV file
+   - Adds a new lakehouse_insert.py script to read CSV file and insert data into IBM Lakehouse tables